Standalone components

+++ src/app/admin/product/product.component.ts
@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component, Inject, PLATFORM_ID } from '@angular/core';
 import { ProductService } from './product.service';
 import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
 import { iProduct } from 'src/app/model/iProduct';
@@ -6,18 +6,23 @@ import { AddEditProductComponent } from '../add-edit-product/add-edit-product.co
 import { Observable } from 'rxjs';
 import { HelperService } from 'src/app/helper/helper.service';
 import { iKategorie } from 'src/app/model/iKategorie';
+import { MatTableModule } from '@angular/material/table';
+import { CommonModule, isPlatformServer } from '@angular/common';
+import { MatButtonModule } from '@angular/material/button';
 
 @Component({
   selector: 'app-product',
   templateUrl: './product.component.html',
   styleUrls: ['./product.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
+  imports: [MatTableModule, CommonModule, MatButtonModule]
 })
 export class ProductComponent {
   displayedColumns: string[] = ['prodid','artid', 'name', 'preis', 'verfugbar', 'edit', 'delete'];
   productsSig = this.prodService.productsSig;
   del$ = new Observable();
-  constructor( private readonly prodService: ProductService, private readonly dialog: MatDialog, private helperService: HelperService) {
+  constructor( private readonly prodService: ProductService, private readonly dialog: MatDialog, private helperService: HelperService, @Inject(PLATFORM_ID) private platformId: any) {
     this.helperService.kategorySig.set({id: 0} as iKategorie);
   }
 
@@ -30,6 +35,9 @@ export class ProductComponent {
     this.dialog.open(AddEditProductComponent, conf);
   }
   deleteProdukt(prod: iProduct) {
+    if(isPlatformServer(this.platformId))
+      return;
+
    const yes = window.confirm('Bist du sicher das du der Produkt '+ prod.name +' löschen willst ?');
    if (yes && prod.id) {
     this.del$ = this.prodService.deleteProduct(prod.id);
